ProfileFragment implements and working
