--- app/src/main/java/com/example/ecostyle/ProfileActivity.kt
@@ -1,46 +1,0 @@
-package com.example.ecostyle
-
-import android.os.Bundle
-import android.widget.Button
-import android.widget.TextView
-import androidx.appcompat.app.AppCompatActivity
-import com.google.firebase.auth.FirebaseAuth
-
-enum class ProviderType {
-    BASIC
-}
-
-class ProfileActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_profile)
-
-        // Inicializar vistas DENTRO de onCreate
-        val emailTextView = findViewById<TextView>(R.id.emailTextView)
-        val providerTextView = findViewById<TextView>(R.id.providerTextView)
-        val logOutButton = findViewById<Button>(R.id.logOutButton)
-
-        val bundle = intent.extras
-        val email = bundle?.getString("email")
-        val provider = bundle?.getString("provider")
-
-        setup(email ?: "", provider ?: "", emailTextView, providerTextView, logOutButton)
-    }
-
-    private fun setup(
-        email: String,
-        provider: String,
-        emailTextView: TextView,
-        providerTextView: TextView,
-        logOutButton: Button
-    ) {
-        title = "Inicio"
-        emailTextView.text = email
-        providerTextView.text = provider
-
-        logOutButton.setOnClickListener {
-            FirebaseAuth.getInstance().signOut()
-            onBackPressed()
-        }
-    }
-}
